requirement.txt
Package                 Version
----------------------- ------------
absl-py                 2.0.0
albumentations          1.3.1
annotated-types         0.6.0
anyio                   4.3.0
appdirs                 1.4.4
black                   24.1.0
cachetools              5.3.2
certifi                 2023.11.17
charset-normalizer      3.3.2
click                   8.1.7
cobra                   0.29.0
colorama                0.4.6
contourpy               1.2.0
cycler                  0.12.1
depinfo                 2.2.0
diskcache               5.6.3
exceptiongroup          1.2.0
filelock                3.9.0
fonttools               4.47.0
fsspec                  2023.4.0
future                  1.0.0
google-auth             2.25.2
google-auth-oauthlib    1.2.0
grpcio                  1.60.0
h11                     0.14.0
httpcore                1.0.4
httpx                   0.27.0
idna                    3.6
imageio                 2.33.1
importlib_resources     6.3.0
Jinja2                  3.1.2
joblib                  1.3.2
kiwisolver              1.4.5
lazy_loader             0.3
lightning-utilities     0.11.2
Markdown                3.5.1
markdown-it-py          3.0.0
MarkupSafe              2.1.3
matplotlib              3.8.2
mdurl                   0.1.2
mpmath                  1.3.0
mypy-extensions         1.0.0
networkx                3.0
numpy                   1.26.2
oauthlib                3.2.2
opencv-python-headless  4.9.0.80
optlang                 1.8.1
packaging               23.2
pandas                  2.1.4
pathspec                0.12.1
Pillow                  9.3.0
pip                     24.0
platformdirs            4.1.0
protobuf                4.23.4
pyasn1                  0.5.1
pyasn1-modules          0.3.0
pydantic                2.6.4
pydantic_core           2.16.3
Pygments                2.17.2
pyparsing               3.1.1
python-dateutil         2.8.2
python-libsbml          5.20.2
pytorch-ssim            0.1
pytz                    2023.3.post1
PyYAML                  6.0.1
qudida                  0.0.4
requests                2.31.0
requests-oauthlib       1.3.1
rich                    13.7.1
rsa                     4.9
ruamel.yaml             0.18.6
ruamel.yaml.clib        0.2.8
scikit-image            0.22.0
scikit-learn            1.3.2
scipy                   1.11.4
setuptools              68.0.0
six                     1.16.0
sniffio                 1.3.1
swiglpk                 5.0.10
sympy                   1.12
tensorboard             2.15.1
tensorboard-data-server 0.7.2
threadpoolctl           3.2.0
tifffile                2023.12.9
tomli                   2.0.1
torch                   2.1.2+cu118
torch-tb-profiler       0.4.3
torchaudio              2.1.2+cu118
torchmetrics            1.3.2
torchvision             0.16.2+cu118
tqdm                    4.66.1
typing_extensions       4.10.0
tzdata                  2023.4
urllib3                 2.1.0
Werkzeug                3.0.1
wheel                   0.41.0
